Remove emails/headderegistry.py
Combine the code with that in emails/policy.py
